Remaining test coverage for TOS page
